requirements.txt
python-dotenv
pandas
tqdm
pillow
openpyxl
requests
